+++ suitkaise/cucumber/__init__.py
@@ -19,6 +19,6 @@ from .api import *
 from .api import __all__
 
 # Module metadata
-__version__ = "0.4.5b0"
+__version__ = "0.4.6b0"
 __author__ = "Suitkaise Development Team"